--- packages/client/@types/Button.d.ts
@@ -1,5 +1,0 @@
-type TButton = {
-  type: 'primary' | 'secondary' | 'accent'
-  children: string
-  clickHandler?: MouseEventHandler | undefined
-}
