--- packages/client/@types/Registration.d.ts
@@ -1,9 +1,0 @@
-declare interface RegFormData {
-  first_name: string
-  second_name: string
-  login: string
-  email: string
-  password: string
-  repassword: string
-  phone: string
-}
