+++ packages/client/src/components/Button/Button.tsx
@@ -1,10 +1,17 @@
+import { MouseEventHandler } from 'react'
 import styles from './Button.module.css'
+
+type ButtonProps = {
+  type: 'primary' | 'secondary' | 'accent'
+  children: string
+  clickHandler?: MouseEventHandler | undefined
+}
 
 export const Button = ({
   type = 'primary',
   children,
   clickHandler,
-}: TButton) => {
+}: ButtonProps) => {
   const className = `${styles.btn} ${styles['btn-' + type]}`
 
   return (
